+++ CHANGELOG.md
@@ -4,12 +4,18 @@
 
 [Unreleased]: https://github.com/chaosiq/chaosiq-cloud/compare/0.5.0...HEAD
 
+### Added
+
+- Generate a `".chaosiq"` workspace file for registering experiment IDs
+  related to their local file names. [#30][30]
+
 ### Changed
 
 - Adapted to use the organization/team schema of the remote API [#27][27]
 - Fixed: features were reset to default value when updating settings [#31][31]
 
 [27]: https://github.com/chaosiq/chaosiq-cloud/issues/27
+[30]: https://github.com/chaosiq/chaosiq-cloud/issues/30
 [31]: https://github.com/chaosiq/chaosiq-cloud/issues/31
 
 ## [0.5.0][] - 2019-11-26